feat: add useCourseFromUrl hook
