+++ src/views/components/calendar/Calendar.tsx
@@ -9,6 +9,7 @@ import ImportantLinks from '@views/components/calendar/ImportantLinks';
 import Divider from '@views/components/common/Divider';
 import CourseCatalogInjectedPopup from '@views/components/injected/CourseCatalogInjectedPopup/CourseCatalogInjectedPopup';
 import { CalendarContext } from '@views/contexts/CalendarContext';
+import useCourseFromUrl from '@views/hooks/useCourseFromUrl';
 import { useFlattenedCourseSchedule } from '@views/hooks/useFlattenedCourseSchedule';
 import { CourseCatalogScraper } from '@views/lib/CourseCatalogScraper';
 import getCourseTableRows from '@views/lib/getCourseTableRows';
@@ -25,21 +26,7 @@ import TeamLinks from './TeamLinks';
 export default function Calendar(): JSX.Element {
     const { courseCells, activeSchedule } = useFlattenedCourseSchedule();
 
-    const [course, setCourse] = useState<Course | null>((): Course | null => {
-        const urlParams = new URLSearchParams(window.location.search);
-        const uniqueIdRaw = urlParams.get('uniqueId');
-        if (uniqueIdRaw === null) return null;
-
-        const uniqueId = Number(uniqueIdRaw);
-        const course = activeSchedule.courses.find(course => course.uniqueId === uniqueId);
-        if (course === undefined) return null;
-
-        urlParams.delete('uniqueId');
-        const newUrl = `${window.location.pathname}?${urlParams}`.replace(/\?$/, '');
-        window.history.replaceState({}, '', newUrl);
-
-        return course;
-    });
+    const [course, setCourse] = useState<Course | null>(useCourseFromUrl());
 
     const [showPopup, setShowPopup] = useState<boolean>(course !== null);
     const [showSidebar, setShowSidebar] = useState<boolean>(true);
@@ -136,6 +123,8 @@ export default function Calendar(): JSX.Element {
                 </div>
 
                 <CourseCatalogInjectedPopup
+                    // Ideally let's not use ! here, but it's fine since we know course is always defined when showPopup is true
+                    // Let's try to refactor this
                     course={course!} // always defined when showPopup is true
                     onClose={() => setShowPopup(false)}
                     open={showPopup}
